examples: add #[allow(unused_attributes)] around "mod"
When examples are compiled independently, they need the enable the
allocator_api feature.  When they are brought in as modules from
other examples, the parent crate does it for them.  In the latter
case, the compiler complains that the module cannot enable the
feature and therefore the attribute is unused.  Follow the example
already present in examples/mutex.rs, and shut up the compiler
about cases that are expected.

Signed-off-by: Paolo Bonzini <pbonzini@redhat.com>